+++ src/aws/index.ts
@@ -1,0 +1,13 @@
+import { accountId } from "./account-id.js";
+import { arn } from "./arn.js";
+import { partition } from "./partition.js";
+import { region } from "./region.js";
+
+const aws = {
+  accountId,
+  arn,
+  region,
+  partition,
+};
+
+export { aws };