+++ test/unit/aws/index.ts
@@ -1,0 +1,13 @@
+import { accountIdTests } from "./account-id.js";
+import { arnTests } from "./arn.js";
+import { partitionTests } from "./partition.js";
+import { regionTests } from "./region.js";
+
+export function injectAwsTests() {
+  describe("aws", () => {
+    accountIdTests();
+    arnTests();
+    partitionTests();
+    regionTests();
+  });
+}