+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2026-01-05
+
+### Added
+
+- Several `aws` related schemas:
+  - `zu.aws.accountId()`, to validate that a string follows the AWS account ID format
+  - `zu.aws.arn()`, to validate that a string is a valid AWS ARN.
+  - `zu.aws.partition()`, to validate that a string is a valid AWS partition.
+  - `zu.aws.region()`, to validate that a string is a valida AWS region.
+
 ## [0.8.0] - 2026-01-02
 
 ### Added
@@ -101,6 +111,7 @@ publication didn't succeed completely. This is a re-release.
 
 - First iteration of the library. It has JSON parsing utilities.
 
+[0.9.0]: https://github.com/infra-blocks/ts-zod-utils/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/infra-blocks/ts-zod-utils/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/infra-blocks/ts-zod-utils/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/infra-blocks/ts-zod-utils/compare/v0.5.0...v0.6.0